+++ test/jest/__mock__/index.js
@@ -1,6 +1,7 @@
 import './currencyData.mock';
 import './documentCreateRange.mock';
 import './matchMedia.mock';
+import './reactVirtializedAutoSizer.mock';
 import './stripesConfig.mock';
 import './stripesIcon.mock';
 import './stripesCore.mock';