+++ src/settings/MappingProfiles/detailsSections/utils.js
@@ -286,12 +286,11 @@ export const getMappingQueryFromValue = valueFromDetails => {
   return valueFromDetails?.substring(0, mappingEndPosition === -1 ? valueFromDetails.length : mappingEndPosition);
 };
 
-export const clearFieldValue = ({ paths, setReferenceTables, isSubfield = false }) => {
-  if (isSubfield) {
-    paths.forEach(path => setReferenceTables(path, []));
-  } else {
-    paths.forEach(path => setReferenceTables(path, ''));
-  }
+export const clearFieldValue = ({ paths, setReferenceTables, isSubfield = false, isCheckbox = false }) => {
+  const checkSubfieldInitialValue = isSubfield ? [] : '';
+  const initialValue = isCheckbox ? BOOLEAN_ACTIONS.ALL_FALSE : checkSubfieldInitialValue;
+
+  paths.forEach(path => setReferenceTables(path, initialValue));
 };
 
 export const clearSubfieldValue = ({