+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ## **6.0.3** (in progress)
 
 ### Bugs fixed:
+* Order field mapping profile: Disable Organization lookup-up when Access provider field disabled (UIDATIMP-1418)
 * Mock `react-virtualized-auto-sizer` module for unit tests (UIDATIMP-1422)
 
 ## [6.0.2](https://github.com/folio-org/ui-data-import/tree/v6.0.2) (2023-03-24)