+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change history for ui-data-import
 
+## **6.0.3** (in progress)
+
+### Bugs fixed:
+* Order field mapping profile: Fix the View UI for Fund value % and Currency (UIDATIMP-1404)
+
 ## [6.0.2](https://github.com/folio-org/ui-data-import/tree/v6.0.2) (2023-03-24)
 
 ### Bugs fixed: